fix(agent.processJWT): handle errors correctly
The way the promise then/catch block was written would cause errors from
processInteractionToken to be handled by the same code that handles
errors from getInteraction. It would result in failures in
processInteractionToken leading to the creation of a new instance of
Interaction.

Also now throws errors which are not NoSuchInteraction (ie. don't eat up
unhandled errors from getInteraction)